requirements.txt
# Requirements para reproducir los experimentos RAG
# Instalar con: pip install -r requirements.txt

# Extracción de texto de PDF
PyMuPDF==1.26.7

# Chunking semántico y embeddings
sentence-transformers==5.2.2

# Cálculo de similitudes
numpy==2.4.2

# API de Google Gemini (LLM)
google-genai==1.63.0
